Refactored by moving source to folders
