+++ avs_hull_web/lib/geometry/hull_math.dart
