+++ avs_hull_web/lib/geometry/point_3d.dart
