+++ avs_hull_web/lib/geometry/spline.dart
